services/worker/requirements.txt
paho-mqtt
requests
numpy